requirements.txt
# Core dependencies for Ragozin Sheets Parser
fastapi>=0.100.0
uvicorn>=0.20.0
streamlit>=1.25.0
python-multipart>=0.0.5

# PDF processing
PyMuPDF>=1.23.0
pdfplumber>=0.9.0
Pillow>=9.0.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# AI and API
openai>=1.0.0
requests>=2.28.0
python-dotenv>=1.0.0

# Visualization
plotly>=5.0.0

# Database (Postgres for production)
psycopg2-binary>=2.9.0

# Auth (Google OAuth + JWT)
authlib>=1.3.0
python-jose[cryptography]>=3.3.0
httpx>=0.25.0
itsdangerous>=2.1.0

# Optional: For traditional pdf2image approach (requires Poppler)
# pdf2image>=3.1.0
